+++ CHANGELOG.md
@@ -28,6 +28,8 @@ BUG FIXES:
       `docker stop`. [GH-3798]
   - providers/docker: Better error messaging when SSH is not ready
       direct to container. [GH-3763]
+  - providers/docker: Don't port map SSH port if container doesn't
+      support SSH. [GH-3857]
   - providers/virtualbox: On Windows, check `VBOX_MSI_INSTALL_PATH`
       for VBoxManage path as well. [GH-3852]
   - provisioners/puppet: Fix setting facter vars with Windows